requirements.txt
django
djangorestframework
gunicorn
python-dotenv
whitenoise
